app/requirements.txt
Django==4.2.13
djangorestframework==3.15.2

# Services tools
psycopg2-binary==2.9.9
gunicorn==23.0.0

# QA
black==25.1.0
isort==6.0.1
Faker==37.1.0
